requirements.txt
fastapi
uvicorn
pillow
python-dotenv
langchain
langchain-google-genai
langchain-groq
langchain-openai
pydantic-settings